Add writeData argument for updating
Currently, `mplusModeler()` calls `prepareMplusData()` to (re)write `rdata` to disk (*.dat) despite that the `mplusObject` has been modified by `update.mplusObject()` and there have been no changes to `rdata`. This results in massive extra time for multiple models using the same large dataset (e.g. imputed data list). I think the easiest implementation (as committed here) is to add a Boolean attribute (`writeData`) in the `mplusObject` that dictates whether or not `prepareMplusData()` will write to disk. `writeData` is by default TRUE, except when `mplusObject` has been modified by `update.mplusObject()` AND `rdata=NULL` (i.e. no changes to rdata). *.inp is written to disk as before. 
There are only 3 real line changes to this file; on lines 1285, 1287, and 1417 (in new file).  The rest are due to indents from the if-statement.

P.S. This is my first commit so apologies if I didn't understand the basics.. Feel free to discard.